feat: convert to EE plugin
BREAKING CHANGE: A valid license is now needed to use this policy
