+++ src/main/java/com/graviteesource/policy/dynamicrouting/DynamicRoutingPolicy.java
@@ -1,29 +1,16 @@
-/**
- * Copyright (C) 2015 The Gravitee team (http://gravitee.io)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.gravitee.policy.dynamicrouting;
+package com.graviteesource.policy.dynamicrouting;
 
+import com.graviteesource.policy.dynamicrouting.configuration.DynamicRoutingPolicyConfiguration;
+import com.graviteesource.policy.dynamicrouting.configuration.Rule;
+import com.graviteesource.policy.dynamicrouting.deployer.DynamicRoutingPolicyDeploymentLifecycle;
 import io.gravitee.common.http.HttpStatusCode;
 import io.gravitee.gateway.api.ExecutionContext;
 import io.gravitee.gateway.api.Request;
 import io.gravitee.gateway.api.Response;
+import io.gravitee.plugin.api.annotations.Plugin;
 import io.gravitee.policy.api.PolicyChain;
 import io.gravitee.policy.api.PolicyResult;
 import io.gravitee.policy.api.annotations.OnRequest;
-import io.gravitee.policy.dynamicrouting.configuration.DynamicRoutingPolicyConfiguration;
-import io.gravitee.policy.dynamicrouting.configuration.Rule;
 import java.io.UnsupportedEncodingException;
 import java.net.URLDecoder;
 import java.nio.charset.Charset;
@@ -42,6 +29,7 @@ import org.slf4j.LoggerFactory;
  * @author Nicolas GERAUD (nicolas.geraud at graviteesource.com)
  * @author GraviteeSource Team
  */
+@Plugin(deployment = DynamicRoutingPolicyDeploymentLifecycle.class)
 public class DynamicRoutingPolicy {
 
     /**
