+++ src/main/resources/plugin.properties
@@ -2,7 +2,7 @@ id=dynamic-routing
 name=Dynamic Routing
 version=${project.version}
 description=${project.description}
-class=io.gravitee.policy.dynamicrouting.DynamicRoutingPolicy
+class=com.graviteesource.policy.dynamicrouting.DynamicRoutingPolicy
 type=policy
 category=others
 icon=dynamic-routing.svg
