+++ src/test/java/com/graviteesource/policy/dynamicrouting/configuration/DynamicRoutingPolicyConfigurationTest.java
@@ -1,19 +1,4 @@
-/**
- * Copyright (C) 2015 The Gravitee team (http://gravitee.io)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.gravitee.policy.dynamicrouting.configuration;
+package com.graviteesource.policy.dynamicrouting.configuration;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.IOException;
@@ -29,7 +14,9 @@ public class DynamicRoutingPolicyConfigurationTest {
 
     @Test
     public void testConfiguration_singleRule() throws IOException {
-        DynamicRoutingPolicyConfiguration configuration = load("/io/gravitee/policy/dynamicrouting/configuration/configuration1.json");
+        DynamicRoutingPolicyConfiguration configuration = load(
+            "/com/graviteesource/policy/dynamicrouting/configuration/configuration1.json"
+        );
 
         Assertions.assertThat(configuration.getRules()).hasSize(1);
     }
