+++ src/test/resources/com/graviteesource/policy/dynamicrouting/configuration/configuration1.json
